Publishing opticalpy package
