+++ pyproject.toml
@@ -1,0 +1,28 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "opticalpy"
+version = "1.0.0"
+authors = [
+    {name="Aurélien Genin", email="aurelien.genin@polytechnique.org"}
+]
+description = "Exact geometrical optics including dispersion to learn and design optical instruments"
+readme = "README.md"
+license = {file = "LICENSE"}
+requires-python = ">=3.10"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
+    "Operating System :: OS Independent",
+]
+dependencies = [
+    "autograd",
+    "matplotlib",
+    "astropy",
+    "scipy"
+]
+
+[project.urls]
+Source = "https://github.com/AstroAure/Opticalpy"
