+++ src/opticalpy/__init__.py
@@ -1,0 +1,6 @@
+from .scene import Scene
+from .ray import Ray
+from .optic import Optic
+from .optgroup import OpticalGroup
+from .utils import wavelength2RGB, dist_angle
+from .visualization import Camera
